feat: health check api 견본

+++ src/main/java/com/dnd8th/service/TaskService.java
@@ -5,10 +5,14 @@ import com.dnd8th.dao.TaskUpdateDao;
 import com.dnd8th.dto.task.TaskCreateRequest;
 import com.dnd8th.entity.Block;
 import com.dnd8th.entity.Task;
+import com.dnd8th.entity.User;
 import com.dnd8th.error.exception.block.BlockNotFoundException;
+import com.dnd8th.error.exception.task.TaskAccessDeniedException;
 import com.dnd8th.error.exception.task.TaskNotFoundException;
+import com.dnd8th.error.exception.user.UserNotFoundException;
 import com.dnd8th.repository.BlockRepository;
 import com.dnd8th.repository.TaskRepository;
+import com.dnd8th.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -23,6 +27,7 @@ public class TaskService {
     private final TaskRepository taskRepository;
     private final BlockRepository blockRepository;
     private final TaskUpdateDao taskUpdateDao;
+    private final UserRepository userRepository;
 
     public Task createTask(TaskCreateRequest taskCreateRequest, Long blockId) {
         Block block = blockRepository.findById(blockId).orElseThrow(BlockNotFoundException::new);
@@ -30,16 +35,37 @@ public class TaskService {
         return taskRepository.save(taskCreateRequest.toEntity(block));
     }
 
-    public void deleteTask(Long taskId) {
+    public void deleteTask(String email, Long taskId) {
+        User user = userRepository.findByEmail(email).orElseThrow(UserNotFoundException::new);
         Task task = taskRepository.findById(taskId).orElseThrow(TaskNotFoundException::new);
+
+        User taskOwner = task.getBlock().getUser();
+        if (taskOwner != user) {
+            throw new TaskAccessDeniedException();
+        }
         taskRepository.delete(task);
     }
 
-    public void updateTask(Long taskId, String content) {
+    public void updateTask(String email, Long taskId, String content) {
+        User user = userRepository.findByEmail(email).orElseThrow(UserNotFoundException::new);
+        Task task = taskRepository.findById(taskId).orElseThrow(TaskNotFoundException::new);
+
+        User taskOwner = task.getBlock().getUser();
+        if (taskOwner != user) {
+            throw new TaskAccessDeniedException();
+        }
+
         taskUpdateDao.updateContent(taskId, content);
     }
 
-    public void toggleStatus(Long taskId) {
+    public void toggleStatus(String email, Long taskId) {
+        User user = userRepository.findByEmail(email).orElseThrow(UserNotFoundException::new);
+        Task task = taskRepository.findById(taskId).orElseThrow(TaskNotFoundException::new);
+
+        User taskOwner = task.getBlock().getUser();
+        if (taskOwner != user) {
+            throw new TaskAccessDeniedException();
+        }
         taskUpdateDao.updateStatus(taskId);
     }
 }
